Updated NewHope to latest from origin

+++ QuantumGateCryptoLib/NewHope/ref/fips202.c
@@ -44,7 +44,7 @@ static void store64(uint8_t *x, uint64_t u)
   unsigned int i;
 
   for(i=0; i<8; ++i) {
-    x[i] = u;
+    x[i] = (uint8_t)u;
     u >>= 8;
   }
 }
